Add a new notification setting: get notifications when an issue a user has commented on, gets updated.
Use case: we do not use teams in TBG (they are too volatile in our work flow)
but there are "areas of interest", so a user might add a comment to an issue
regarding implementation etc. In the past it has occured that the assigned user
then replied (e.g. with additional questions), and the commenter was never notified.

Following an issue might be used as well, but there are usually too many issues
for this to be used.