Update actions.class.php to support GitLab webhook format.
GitLab does not wrap the request in a 'payload' parameter. Since there is no parameter to obtain, you can check and see if the raw request contains the required information. If so, you can proceed normally. This will allow TBG to accept wenhook data from GitLab and use the current Github functionality until such time as a GitLab option is added. No other changes seem to be required to allow GitLab to utilize this functionality.