Improved sidebar visuals
Improved the sidebar collapser link visuals (all issues were seen in the
issues overview view with the "Predefined searches" sidebar):

- The collapser link on the sidebar was positioned absolute with a fix
distance of pixels. This caused a problem when the browser was not in
fullscreen mode and the window was resized

- The collapser link was partly covered (or even completely without the
fix above) by the header "Predefined searches" whose z-Order seemed to
be greater.