Duplicate Edition IDs In Use on Release Center
Instead of showing edition builds under "Project Releases" and "Edition
Releases", this commit only shows them under the appropriate Edition
section.  This prevents the issue of having duplicate IDs on the page.

This commit also fixes a bug where TBGBuild::isEditionBuild() was
incorrectly identifying builds as being associated with an addition.
The edition ID was 0 instead of null.

Finally, some of the markup was modified so that when the final build
was removed from any section, the muted text informing the user that
there are no releases for the project / edition is restored.

This issue fixes #2459