"Create new users" hide dialog when user were successfully created
When creating a new user by the "More details" dialog, the dialog
(backdrop) stays open, event when the user were created successfully.
This is very confusing because most of the time the user expects that
the backdrop closes after a successful creation. This little feature has
been added with this small fix.