Change style for "Configure issue types" and "Issuetype scheme" proposal
When configuring an issuetype scheme, the dropdown box for the
"Redirect" option was always larger than the box it was embedded into.
Since word wrap is not possible for plain HTML select boxes the whole
"Configure issue types" style needed to be adapted a little bit. It is
now 50 pixel larger than the other tabs (which no one notices) and the
size of the select boxes (including padding) were adapted.