Fixed issues with functionalities inside "Manage issue subscribers" modal.
On action "Add a user", instead of closing dropdown when user is selected, ajax loading indicator inside dropdown is added to be consistent with other alike issue modals. Subscribing to issue via star has visually stayed the same but javascript logic was adapted to support toggle of aforementioned new indicator.