Improved UX of showing / hiding sticky issues paginator.
Problem for UX with previous solution was that if you scroll to the point when element changes "position" from fixed to relative or vice versa content jumping will happen. With this solution that is no longer the issue since there are already two ".issue_paginator" elements and only the fixed one will hide / show itself. Additionally this commit can be improved by duplicating ".issue_paginator" element via javascript and not html.