Additional fixes for roadmap milestone / sprint burndown graph calculation.
Before this fix (and before commit 1c120636563344db2412374776e6c1aed80d6551) calculation was still relying on multi day estimations even for same issue which is now changed not to be daily and only use last issue estimation. Spent time hours and points are unchanged ie still additive.

New change is rising up first (highest) burndown dot for the difference between spent and estimated when spent exceedes estimated so that burndown line is visible and does not continue on y minus axis which is not visible.