Fixed position of caret when mentioning users.
Caret will now be moved to the end of mentioned user string instead of end of whole text. This is achieved by storing caret position before and restoring after text update. Since jQuery plugins"mention" and "bootstrap-typeahead" are old, and there aren't updates, their source was altered to accomplish this.