Require issue add permission for adding epics
Since adding an epic adds an issue to the board, it seems that the user should need to have permission to add an issue if they want to add an epic.  Right now it only requires permission to view the board.  A user might have view permission but not issue add permission, such as the case of a guest user.

This copies the issue add permission check into the add epic function in the agile module.

Note that it currently doesn't display an error message, it just silently fails (this would be a good improvement in future).