Require higher level permissions for agile changes
Currently, several operations that permit viewing agile boards also allow them to be modified.

It would be nice for operations that change agile boards to be restricted to users with a higher permission set.  IE, it would be useful to be able to show agile boards to guest users, without also allowing guest users to modify them.

This change expands the existing 'project_scrum' permission, which was previously only used to gate assigning milestones, to also gate several other modification checks.  (It seemed like this was intended to be like a scrum master type permission (?))

Note that it currently doesn't display an error message, or hide the edit buttons -- it just silently fails (this would be a good improvement in future).